Moved ServerHandler helper class to tests.
It has been only used in the builtin_servers tests since Django 1.4.
